chore: move to GH
The change (originally in Gerrit/OpenStackSDK) is becoming too big and
makes sense to be separated into a standalone project. This is however
requires administrative work with respect to requesting TC approval ...
So For now just move development to the GitHub and move it back to
opendev once administrative work is complete.