feat: introduce type model
in order to allow code generation for different languages it is required
to have a universal data model. Using json schema directly is not very
friendly and leads to huge amount of code duplication (as previous code
shows). With this change an abstract DataModel is being introduced which
is then used by the specific TypeManagers to get types suitable for the
target.

Reorganization is not completed with that change, but we reached some
milestone allowing further work in smaller chunks.