feat(keystone): improve auth schema generation
Real schema in the kestone code is useless since it only describes
password or token based authorization. Replace it with the proper schema
that also now contains MFA authorization.
Make rust structure fields public (in crate) since in some cases it is
required in the OSC to have access to the filled data of the
RestEndpoint with no access to the raw data.