feat: use api version in the openapi file name
use api version (major/minor) in the file name of the generated openapi
schema and create a symbolic link to the latest one. This allows proper
versioning.
In the neutron case use neutron major version as api minor version:
neutron 23.1.0 => api version 2.23. This is not the best way but there
is no reasonable alternative